Rename: Entry -> EntryDesktop
