+++ src/components/entry/EntryOverview.js
@@ -20,7 +20,7 @@ import { ReportModal } from '../form/ReportModal'
 import Button from '../ui/Button'
 import { theme } from '../ui/GlobalStyle'
 import IconBesideText from '../ui/IconBesideText'
-import { TextContent } from './Entry'
+import { TextContent } from './EntryDesktop'
 import EntryTags from './EntryTags'
 import { ReviewButton } from './ReviewButton'
 import { formatISOString, formatSeasonality } from './textFormatters'
