+++ src/components/entry/EntryReviews.js
@@ -4,7 +4,7 @@ import { openLightbox } from '../../redux/locationSlice'
 import { useAppHistory } from '../../utils/useAppHistory'
 import { useIsDesktop } from '../../utils/useBreakpoint'
 import { ReviewForm } from '../form/ReviewForm'
-import { TextContent } from './Entry'
+import { TextContent } from './EntryDesktop'
 import Review from './Review'
 import { ReviewButton } from './ReviewButton'
 import ReviewSummary from './ReviewSummary'
